requirements.txt
flask>=3.0
google-api-python-client>=2.100
google-auth-oauthlib>=1.2
google-auth>=2.25
requests>=2.31
